+++ src/sdk/PnP.Core.Auth/Confidential/OnBehalfOfAuthenticationProvider.cs
@@ -36,6 +36,11 @@ namespace PnP.Core.Auth
         /// </summary>
         public Func<string> UserTokenProvider { get; set; }
 
+        /// <summary>
+        /// Share the cache between all ClientApplication objects. The cache becomes static. Defaults to false.
+        /// </summary>
+        public bool UseSharedCache { get; set; }
+
         // Instance private member, to keep the token cache at service instance level
         private IConfidentialClientApplication confidentialClientApplication;
 
@@ -49,13 +54,15 @@ namespace PnP.Core.Auth
         /// <param name="tenantId">The Tenant ID for the Authentication Provider</param>
         /// <param name="clientSecret">The Client Secret of the app</param>
         /// <param name="userTokenProvider">A function providing the consumer user access token to use for the On-Behalf-Of flow</param>
+        /// <param name="useSharedCache">Share the cache between all ClientApplication objects. The cache becomes static. Defaults to false</param>
         public OnBehalfOfAuthenticationProvider(string clientId, string tenantId,
             SecureString clientSecret,
-            Func<string> userTokenProvider)
+            Func<string> userTokenProvider,
+            bool useSharedCache = false)
             : this(clientId, tenantId, new PnPCoreAuthenticationOnBehalfOfOptions
             {
                 ClientSecret = clientSecret?.ToInsecureString()
-            }, userTokenProvider)
+            }, userTokenProvider, useSharedCache)
         {
         }
 
@@ -66,12 +73,15 @@ namespace PnP.Core.Auth
         /// <param name="tenantId">The Tenant ID for the Authentication Provider</param>
         /// <param name="options">Options for the authentication provider</param>
         /// <param name="userTokenProvider">A function providing the consumer user access token to use for the On-Behalf-Of flow</param>
+        /// <param name="useSharedCache">Share the cache between all ClientApplication objects. The cache becomes static. Defaults to false</param>
         public OnBehalfOfAuthenticationProvider(string clientId, string tenantId,
             PnPCoreAuthenticationOnBehalfOfOptions options,
-            Func<string> userTokenProvider)
+            Func<string> userTokenProvider,
+            bool useSharedCache = false)
             : this(null, null)
         {
             UserTokenProvider = userTokenProvider;
+            UseSharedCache = useSharedCache;
             Init(new PnPCoreAuthenticationCredentialConfigurationOptions
             {
                 ClientId = clientId,
@@ -89,12 +99,15 @@ namespace PnP.Core.Auth
         /// <param name="storeLocation">The Store Location to get the X.509 certificate from</param>
         /// <param name="thumbprint">The Thumbprint of the X.509 certificate</param>
         /// <param name="userTokenProvider">A function providing the consumer user access token to use for the On-Behalf-Of flow</param>
+        /// <param name="useSharedCache">Share the cache between all ClientApplication objects. The cache becomes static. Defaults to false</param>
         public OnBehalfOfAuthenticationProvider(string clientId, string tenantId,
             StoreName storeName, StoreLocation storeLocation, string thumbprint,
-            Func<string> userTokenProvider)
+            Func<string> userTokenProvider,
+            bool useSharedCache = false)
             : this(null, null)
         {
             UserTokenProvider = userTokenProvider;
+            UseSharedCache = useSharedCache;
             Init(new PnPCoreAuthenticationCredentialConfigurationOptions
             {
                 ClientId = clientId,
@@ -115,12 +128,15 @@ namespace PnP.Core.Auth
         /// <param name="tenantId">The Tenant ID for the Authentication Provider</param>
         /// <param name="certificate">The X.509 certificate to use for authentication</param>
         /// <param name="userTokenProvider">A function providing the consumer user access token to use for the On-Behalf-Of flow</param>
+        /// <param name="useSharedCache">Share the cache between all ClientApplication objects. The cache becomes static. Defaults to false</param>
         public OnBehalfOfAuthenticationProvider(string clientId, string tenantId,
             X509Certificate2 certificate,
-            Func<string> userTokenProvider)
+            Func<string> userTokenProvider,
+            bool useSharedCache = false)
             : this(null, null)
         {
             UserTokenProvider = userTokenProvider;
+            UseSharedCache = useSharedCache;
             Init(new PnPCoreAuthenticationCredentialConfigurationOptions
             {
                 ClientId = clientId,
@@ -184,9 +200,10 @@ namespace PnP.Core.Auth
                 ClientSecret = options.OnBehalfOf.ClientSecret.ToSecureString();
             }
 
+            ConfidentialClientApplicationBuilder confidentialClientApplicationBuilder;
             if (Certificate != null)
             {
-                confidentialClientApplication = ConfidentialClientApplicationBuilder
+                confidentialClientApplicationBuilder = ConfidentialClientApplicationBuilder
                     .Create(ClientId)
                     .WithCertificate(Certificate)
                     .WithHttpClientFactory(msalHttpClientFactory)
@@ -195,12 +212,11 @@ namespace PnP.Core.Auth
                         options.OnBehalfOf.RedirectUri,
                         TenantId,
                         options.Environment,
-                        options.AzureADLoginAuthority)
-                    .Build();
+                        options.AzureADLoginAuthority);
             }
             else
             {
-                confidentialClientApplication = ConfidentialClientApplicationBuilder
+                confidentialClientApplicationBuilder = ConfidentialClientApplicationBuilder
                     .Create(ClientId)
                     .WithClientSecret(ClientSecret.ToInsecureString())
                     .WithHttpClientFactory(msalHttpClientFactory)
@@ -209,9 +225,18 @@ namespace PnP.Core.Auth
                         options.OnBehalfOf.RedirectUri,
                         TenantId,
                         options.Environment,
-                        options.AzureADLoginAuthority)
-                    .Build();
+                        options.AzureADLoginAuthority);
             }
+
+            if (UseSharedCache)
+            {
+                confidentialClientApplicationBuilder.WithCacheOptions(new CacheOptions()
+                {
+                    UseSharedCache = true
+                });
+            }
+
+            confidentialClientApplication = confidentialClientApplicationBuilder.Build();
 
             // Log the initialization information
             Log?.LogInformation(PnPCoreAuthResources.OnBehalfOfAuthenticationProvider_LogInit);