+++ src/sdk/CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Added xxxClientSideComponentId and xxxFormClientSideComponentProperties properties for `IContentType` #1421 [jansenbe - Bert Jansen]
 - Allow access to SharePoint and Microsoft Graph response headers for interactive/batch requests, including the `SPRequestGuid` (=correlation id) and Microsoft Graph `request-id` [jansenbe - Bert Jansen]
 - Allow access to the SPRequest Guid via the returned `BatchRequest` #1464 [StefanSchoof - Stefan Schoof]
+- Support for `UseSharedCache` in the `OnBehalfOfAuthenticationProvider` #1380 [jansenbe - Bert Jansen]
 
 ### Changed